+++ solc/CommandLineInterface.cpp
@@ -845,7 +845,7 @@ void CommandLineInterface::assembleFromEVMAssemblyJSON()
 	solAssert(m_fileReader.sourceUnits().size() == 1);
 	auto&& [sourceUnitName, source] = *m_fileReader.sourceUnits().begin();
 
-	auto evmAssemblyStack = std::make_unique<evmasm::EVMAssemblyStack>(m_options.output.evmVersion);
+	auto evmAssemblyStack = std::make_unique<evmasm::EVMAssemblyStack>(m_options.output.evmVersion, m_options.output.eofVersion);
 	try
 	{
 		evmAssemblyStack->parseAndAnalyze(sourceUnitName, source);
@@ -891,15 +891,17 @@ void CommandLineInterface::compile()
 		if (m_options.output.debugInfoSelection.has_value())
 			m_compiler->selectDebugInfo(m_options.output.debugInfoSelection.value());
 
-		CompilerStack::IROutputSelection irOutputSelection = CompilerStack::IROutputSelection::None;
-		if (m_options.compiler.outputs.irOptimized || m_options.compiler.outputs.irOptimizedAstJson || m_options.compiler.outputs.yulCFGJson)
-			irOutputSelection = CompilerStack::IROutputSelection::UnoptimizedAndOptimized;
-		else if (m_options.compiler.outputs.ir || m_options.compiler.outputs.irAstJson)
-			irOutputSelection = CompilerStack::IROutputSelection::UnoptimizedOnly;
-
-		m_compiler->requestIROutputs(irOutputSelection);
+		CompilerStack::PipelineConfig pipelineConfig;
+		pipelineConfig.irOptimization =
+			m_options.compiler.outputs.irOptimized ||
+			m_options.compiler.outputs.irOptimizedAstJson ||
+			m_options.compiler.outputs.yulCFGJson;
+		pipelineConfig.irCodegen =
+			pipelineConfig.irOptimization ||
+			m_options.compiler.outputs.ir ||
+			m_options.compiler.outputs.irAstJson;
 		m_compiler->setMLIRGenJobSpec(m_options.mlirGenJob);
-		m_compiler->enableEvmBytecodeGeneration(
+		pipelineConfig.bytecode =
 			m_options.compiler.estimateGas ||
 			m_options.compiler.outputs.asm_ ||
 			m_options.compiler.outputs.asmJson ||
@@ -917,8 +919,9 @@ void CommandLineInterface::compile()
 				m_options.compiler.combinedJsonRequests->srcMapRuntime ||
 				m_options.compiler.combinedJsonRequests->funDebug ||
 				m_options.compiler.combinedJsonRequests->funDebugRuntime
-			))
-		);
+			));
+
+		m_compiler->selectContracts({{"", {{"", pipelineConfig}}}});
 
 		m_compiler->setOptimiserSettings(m_options.optimiserSettings());
 
@@ -1250,7 +1253,8 @@ void CommandLineInterface::assembleYul(yul::YulStack::Language _language, yul::Y
 			{
 			case yul::YulStack::Language::Assembly:
 			case yul::YulStack::Language::StrictAssembly:
-				dialect = &yul::EVMDialect::strictAssemblyForEVMObjects(m_options.output.evmVersion);
+				dialect = &yul::EVMDialect::
+							  strictAssemblyForEVMObjects(m_options.output.evmVersion, m_options.output.eofVersion);
 				break;
 			default:
 				solUnimplementedAssert(false, "Invalid yul dialect");